% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DictionaryGI}
\alias{DictionaryGI}
\title{Dictionary with opinionated words from the Harvard-IV dictionary as used in 
the General Inquirer software}
\format{A list with different terms according to Henry}
\source{
\url{http://www.wjh.harvard.edu/~inquirer/}
}
\usage{
data(DictionaryGI)
}
\description{
Dictionary with a list of positive and negative words according to the psychological 
Harvard-IV dictionary as used in the General Inquirer software. This is a 
general-purpose dictionary developed by the Harvard University.
}
\note{
All words are in lower case and non-stemmed
}
\examples{
data(DictionaryGI)
summary(DictionaryGI)
}
\keyword{datasets}

