% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DictionaryHE}
\alias{DictionaryHE}
\title{Dictionary with opinionated words from Henry's Financial dictionary}
\format{A list with different wordlists according to Henry}
\usage{
data(DictionaryHE)
}
\description{
Dictionary with a list of positive and negative words according to the Henry's 
finance-specific dictionary. This dictionary was first presented in the \emph{Journal 
of Business Communication} among one of the early adopters of text analysis in the 
finance discipline.
}
\note{
All words are in lower case and non-stemmed
}
\examples{
data(DictionaryHE)
summary(DictionaryHE)
}
\references{
Henry (2008): \emph{Are Investors Influenced By How Earnings Press 
Releases Are Written?}, Journal of Business Communication, 45:4, 363-407
}
\keyword{datasets}
