% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DictionaryLM}
\alias{DictionaryLM}
\title{Dictionary with opinionated words from Loughran-McDonald Financial dictionary}
\format{A list with different terms according to Loughran-McDonald}
\source{
\url{http://www3.nd.edu/~mcdonald/Word_Lists.html}
}
\usage{
data(DictionaryLM)
}
\description{
Dictionary with a list of positive, negative and uncertainty words according to the 
Loughran-McDonald finance-specific dictionary. This dictionary was first presented 
in the \emph{Journal of Finance} and has been widely used in the finance domain ever 
since.
}
\note{
All words are in lower case and non-stemmed
}
\examples{
data(DictionaryLM)
summary(DictionaryLM)
}
\references{
Loughran and McDonald (2011) \emph{When is a Liability not a Liability? 
Textual Analysis, Dictionaries, and 10-Ks}, Journal of Finance, 66:1, 35-65
}
\keyword{datasets}

