% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentanalysis.R
\name{getSentimentAnalysis}
\alias{getSentimentAnalysis}
\title{Set options as stringAsFactors = FALSE}
\usage{
getSentimentAnalysis(tweets.df, time.break = "day")
}
\description{
Set options as stringAsFactors = FALSE
}
