% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentanalysis.R
\name{getSentimentAnalysisDF}
\alias{getSentimentAnalysisDF}
\title{function to apply trained model to tweets data.frame}
\usage{
getSentimentAnalysisDF(df_tweets, glmnet_classifier = NULL)
}
\arguments{
\item{df_tweets}{data.frame of tweet comments}

\item{glmnet_classifier}{tained glm_net model}
}
\value{
sentiment_df sentiment analyzed data.frame
}
\description{
function to apply trained model to tweets data.frame
}
