% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentanalysis.R
\name{getSentimentAnalysisPlot}
\alias{getSentimentAnalysisPlot}
\title{function to get plot for sentiment analysis data.frame}
\usage{
getSentimentAnalysisPlot(df_tweets)
}
\arguments{
\item{df_tweets}{data.frame of glm_net calssified tweets}
}
\value{
sentimentplot
}
\description{
function to get plot for sentiment analysis data.frame
}
