% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEstimation.R
\name{glmEstimation}
\alias{glmEstimation}
\title{Estimation via generalized least squares}
\usage{
glmEstimation(x, response, control = list(family = "gaussian"), ...)
}
\arguments{
\item{x}{An object of type \code{\link[tm]{DocumentTermMatrix}}.}

\item{response}{Response variable including the given gold standard.}

\item{control}{(optional) A list of parameters defining the model as follows:
\itemize{
 \item{"family"}{Distribution for response variable. Default is \code{family="gaussian"}.
For non-negative counts, use \code{family="poisson"}. For binary variables
\code{family="binomial"}. See \code{\link{glm}} for further details.}
}}

\item{...}{Additional parameters passed to function for \code{\link{glm}}.}
}
\value{
Result is a list with coefficients, coefficient names and the model intercept.

Result is a list with coefficients, coefficient names and the model intercept.
}
\description{
Function estimates coefficients based on generalized least squares.
}
