% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEstimation.R
\name{lmEstimation}
\alias{lmEstimation}
\title{Ordinary least squares estimation}
\usage{
lmEstimation(x, response, control = list(), ...)
}
\arguments{
\item{x}{An object of type \code{\link[tm]{DocumentTermMatrix}}.}

\item{response}{Response variable including the given gold standard.}

\item{control}{(optional) A list of parameters (not used).}

\item{...}{Additional parameters (not used).}
}
\value{
Result is a list with coefficients, coefficient names and the model intercept.
}
\description{
Function estimates coefficients based on ordinary least squares.
}
