% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{loadDictionaryGI}
\alias{loadDictionaryGI}
\title{Loads Harvard-IV dictionary into object}
\usage{
loadDictionaryGI()
}
\value{
object of class \code{\link{SentimentDictionary}}
}
\description{
Loads Harvard-IV dictionary (as used in General Inquirer) into a standardized
dictionary object
}
\note{
Result is a list of stemmed words in lower case
}

