% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{loadDictionaryHE}
\alias{loadDictionaryHE}
\title{Loads Henry's finance-specific dictionary into object}
\usage{
loadDictionaryHE()
}
\value{
object of class \code{\link{SentimentDictionary}}
}
\description{
Loads Henry's finance-specific dictionary into a standardized dictionary object
}
\note{
Result is a list of stemmed words in lower case
}
