% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{loadDictionaryLM}
\alias{loadDictionaryLM}
\title{Loads Loughran-McDonald dictionary into object}
\usage{
loadDictionaryLM()
}
\value{
object of class \code{\link{SentimentDictionary}}
}
\description{
Loads Loughran-McDonald financial dictionary into a standardized dictionary 
object (here, categories positive and negative are considered)
}
\note{
Result is a list of stemmed words in lower case
}
