% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{loadDictionaryLM_Uncertainty}
\alias{loadDictionaryLM_Uncertainty}
\title{Loads uncertainty words from Loughran-McDonald into object}
\usage{
loadDictionaryLM_Uncertainty()
}
\value{
object of class \code{\link{SentimentDictionary}}
}
\description{
Loads uncertainty words from Loughran-McDonald into a standardized
dictionary object
}
\note{
Result is a list of stemmed words in lower case
}
