% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{loadDictionaryQDAP}
\alias{loadDictionaryQDAP}
\title{Loads polarity words from qdap package into object}
\source{
\url{https://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html}
}
\usage{
loadDictionaryQDAP()
}
\value{
object of class \code{\link{SentimentDictionary}}
}
\description{
Loads polarity words from data object \code{\link[qdapDictionaries]{key.pol}} 
which is by the package qdap. This is then converted into a standardized
dictionary object
}
\note{
Result is a list of stemmed words in lower case
}
\references{
Hu and Liu (2004). Mining Opinion Features in Customer Reviews. 
National Conference on Artificial Intelligence.
}

