% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEstimation.R
\name{lookupEstimationMethod}
\alias{lookupEstimationMethod}
\title{Estimation method}
\usage{
lookupEstimationMethod(type)
}
\arguments{
\item{type}{A string denoting the estimation method. Allowed values are \code{lasso} or \code{spikeslab}.}
}
\value{
Function that implements the specific estimation method.
}
\description{
Decides upon a estimation method for dictionary generation. Input is a name for the estimation method, output is the corresponding function object.
}

