% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\description{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
- cols:   Number of columns in layout
- layout: A matrix specifying the layout. If present, 'cols' is ignored.
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\examples{
x <- 1:10
y1 <- 1:10
y2 <- 21:30

plt1 <- ggplot(data=data.frame(x=x, y=y1)) + geom_point(aes(x,y))
plt2 <- ggplot(data=data.frame(x=x, y=y1)) + geom_line(aes(x,y))
plt3 <- ggplot(data=data.frame(x=x, y=y1 + rnorm(10))) + geom_line(aes(x,y))

multiplot(plt1,plt2,plt3, cols = 2)
}
