% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentanalysis.R
\name{prepareTextForSentimentAnalysis}
\alias{prepareTextForSentimentAnalysis}
\title{Function to prepare text for sentiment analysis
removing special characters, punctuation, numbers links etc..}
\usage{
prepareTextForSentimentAnalysis(textvector)
}
\arguments{
\item{textvector}{vector of text to perform sentiment analysis on}
}
\value{
textvector cleaned text vector
}
\description{
Function to prepare text for sentiment analysis
removing special characters, punctuation, numbers links etc..
}
