% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessCorpus.R
\name{preprocessCorpus}
\alias{preprocessCorpus}
\title{Default preprocessing of corpus to generate a document-term matrix}
\usage{
preprocessCorpus(corpus, language = "english", stemming = TRUE,
  verbose = FALSE)
}
\arguments{
\item{corpus}{\code{\link[tm]{Corpus}} object which should be processed}

\item{language}{default language used for preprocessing (i.e. stop word removal and stemming)}

\item{stemming}{perform stemming (default: TRUE)}

\item{verbose}{print preprocessing status information}
}
\description{
Preprocess existing corpus of type \code{\link[tm]{Corpus}} according to default operations. 
This helper function groups all standard preprocessing steps such that the usage of the 
package is more convenient. The result is a document-term matrix.
}
\seealso{
\code{\link[tm]{DocumentTermMatrix}} and \code{\link[tm]{TermDocumentMatrix}} for
the underlying class
}
\keyword{corpus}
\keyword{preprocessing}

