% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{quantFormatting}
\alias{quantFormatting}
\title{Return the Quant Report header text and header background colours.}
\usage{
quantFormatting()
}
\description{
This function takes no arguments and returns a \emph{list} of 2 Hex codes
describing the colours used in the Quant header.
}
\examples{
\dontrun{
quantheader <- quantFormatting()
scales::show_col(unlist(quantheader))
}
}
