% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentRules.R
\name{ruleLinearModel}
\alias{ruleLinearModel}
\title{Sentiment based on linear model}
\usage{
ruleLinearModel(dtm, d)
}
\arguments{
\item{dtm}{Document-term matrix}

\item{d}{Dictionary of type \code{\link{SentimentDictionaryWeighted}}}
}
\value{
Continuous sentiment score
}
\description{
Sentiment score as denoted by a linear model.
}
\keyword{rules}

