% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentRules.R
\name{ruleNegativity}
\alias{ruleNegativity}
\title{Ratio of negative words}
\usage{
ruleNegativity(dtm, d)
}
\arguments{
\item{dtm}{Document-term matrix}

\item{d}{Dictionary of type \code{\link{SentimentDictionaryBinary}}}
}
\value{
Ratio of negative words compared to all
}
\description{
Ratio of words labeled as negative in that dictionary compared to the total 
number of words in the document. Here, it uses the entry \code{negativeWords}
of the \code{\link{SentimentDictionaryBinary}}.
}
\keyword{rules}
