% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentRules.R
\name{ruleSentiment}
\alias{ruleSentiment}
\title{Sentiment score}
\usage{
ruleSentiment(dtm, d)
}
\arguments{
\item{dtm}{Document-term matrix}

\item{d}{Dictionary of type \code{\link{SentimentDictionaryBinary}}}
}
\value{
Sentiment score in the range of -1 to 1.
}
\description{
Sentiment score defined as the difference between positive and negative
word counts divided by the total number of words.
}
\details{
Given the number of positive words \eqn{P} and the number of 
negative words \eqn{N}. Further, let \eqn{T} denote the total number of words
in that document. Then, the sentiment ratio is defined as 
\deqn{\frac{P+N}{T}}. Here, it uses the entries \code{negativeWords} and
\code{positiveWords} of the \code{\link{SentimentDictionaryBinary}}.
}
\keyword{rules}
