% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentRules.R
\name{ruleSentimentPolarity}
\alias{ruleSentimentPolarity}
\title{Sentiment polarity score}
\usage{
ruleSentimentPolarity(dtm, d)
}
\arguments{
\item{dtm}{Document-term matrix}

\item{d}{Dictionary of type \code{\link{SentimentDictionaryBinary}}}
}
\value{
Sentiment score in the range of -1 to 1.
}
\description{
Sentiment score defined as the difference between positive and negative
word counts divided by the sum of positive and negative words.
}
\details{
Given the number of positive words \eqn{P} and the number of 
negative words \eqn{N}. Then, the sentiment ratio is defined as 
\deqn{\frac{P-N}{P+N}}. Here, it uses the entries \code{negativeWords} and
\code{positiveWords} of the \code{\link{SentimentDictionaryBinary}}.
}
\keyword{rules}
