% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentRules.R
\name{ruleWordCount}
\alias{ruleWordCount}
\title{Counts word frequencies}
\usage{
ruleWordCount(dtm)
}
\arguments{
\item{dtm}{Document-term matrix}
}
\value{
Total number of words
}
\description{
Counts total word frequencies in each document
}
\keyword{rules}

