% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentimentAnalysis.R
\docType{package}
\name{SentimentAnalysis}
\alias{SentimentAnalysis}
\title{SentimentAnalysis: A package for analyzing sentiment of texts}
\description{
The \code{SentimentAnalysis} package provides routines to quickly measure
the sentiment of written materials. It ships a dedicated class
SentimentDictionary to store different variants of dictionaries
(including pre-built ones that are ready to go) and helps the
user with routines for constructing domain-specific dictionaries and
evaluating the performance of common rules for analyzing sentiment.
}
