% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentimentDictionary.R
\name{SentimentDictionary}
\alias{SentimentDictionary}
\title{Create new sentiment dictionary based on input}
\usage{
SentimentDictionary(...)
}
\arguments{
\item{...}{Arguments as passed to one of the three functions 
\code{\link{SentimentDictionaryWordlist}}, \code{\link{SentimentDictionaryBinary}} or
\code{\link{SentimentDictionaryWeighted}}}
}
\description{
Depending on the input, this function creates a new sentiment dictionary of different type.
}
\seealso{
\code{\link{SentimentDictionaryWordlist}}, 
\code{\link{SentimentDictionaryBinary}}, 
\code{\link{SentimentDictionaryWeighted}}
}
\keyword{dictionary}
