% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentimentDictionary.R
\name{SentimentDictionaryBinary}
\alias{SentimentDictionaryBinary}
\title{Create a sentiment dictionary of positive and negative words}
\usage{
SentimentDictionaryBinary(positiveWords, negativeWords)
}
\arguments{
\item{positiveWords}{is a vector containing the entries labeled as positive}

\item{negativeWords}{is a vector containing the entries labeled as negative}
}
\value{
Returns a new object of type \code{SentimentDictionaryBinary}
}
\description{
This routines creates a new object of type \code{SentimentDictionaryBinary} that
stores two separate vectors of negative and positive words
}
\examples{
# generate a dictionary with positive and negative words
d <- SentimentDictionaryBinary(c("increase", "rise", "more"),
                               c("fall", "drop"))
summary(d)
# alternative call
d <- SentimentDictionary(c("increase", "rise", "more"),
                         c("fall", "drop"))
summary(d)
}
\seealso{
\code{\link{SentimentDictionary}}
}
\keyword{dictionary}

