% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentimentDictionary.R
\name{SentimentDictionaryWordlist}
\alias{SentimentDictionaryWordlist}
\title{Create a sentiment dictionary consisting of a simple wordlist}
\usage{
SentimentDictionaryWordlist(wordlist)
}
\arguments{
\item{wordlist}{is a vector containing the individual entries as strings}
}
\value{
Returns a new object of type \code{SentimentDictionaryWordlist}
}
\description{
This routine creates a new object of type \code{SentimentDictionaryWordlist}
}
\examples{
# generate a dictionary with "uncertainty" words
d <- SentimentDictionaryWordlist(c("uncertain", "possible", "likely"))
summary(d)
# alternative call
d <- SentimentDictionary(c("uncertain", "possible", "likely"))
summary(d)
}
\seealso{
\code{\link{SentimentDictionary}}
}
\keyword{dictionary}
