% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{spiderPlotly}
\alias{spiderPlotly}
\title{Create a Spider graph Plotly object}
\usage{
spiderPlotly(dataset, benchmark = mean)
}
\arguments{
\item{dataset}{An input dataset to display in the plot}

\item{benchmark}{The function used to create the benchmark category in the plot (Eg: mean, median, other bespoke function)}
}
\value{
A plotly object
}
\description{
Generic function to create an interactive Spider graph in Plotly
}
\details{
The input dataset needs to have:
\itemize{
 \item First column must be a vector of characters representing the names of categories to be compared
 \item The remaining columns must only contain numeric values that can be converted into spider graph coordinates
}
}
\examples{
# Transforming some data into the format required with the first column containing names
test <- mtcars[1:5,1:7]
test$name <- rownames(test)
test <- test[,c(8,1:7)]

spiderPlotly(test)

}
