% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEstimation.R
\name{spikeslabEstimation}
\alias{spikeslabEstimation}
\title{Spike-and-slab estimation}
\usage{
spikeslabEstimation(
  x,
  response,
  control = list(n.iter1 = 500, n.iter2 = 500),
  ...
)
}
\arguments{
\item{x}{An object of type \code{\link[tm]{DocumentTermMatrix}}.}

\item{response}{Response variable including the given gold standard.}

\item{control}{(optional) A list of parameters defining the LASSO model. Default is\code{n.iter1=500} and \code{n.iter2=500}.
See \code{\link[spikeslab]{spikeslab}} for details.}

\item{...}{Additional parameters passed to function for \code{\link[spikeslab]{spikeslab}}.}
}
\value{
Result is a list with coefficients, coefficient names and the model intercept.
}
\description{
Function estimates coefficients based on spike-and-slab regression.
}
