% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentanalysis.R
\name{trainModelForSentimentAnalysis}
\alias{trainModelForSentimentAnalysis}
\title{function to train model for sentiment clasification}
\usage{
trainModelForSentimentAnalysis(tweets_classified = NULL)
}
\arguments{
\item{textvector}{a text vetor with clened text data}
}
\value{
trained.model glmnet classified model
}
\description{
function to train model for sentiment clasification
}
